--- src/components/expire/Expire.module.css
@@ -1,12 +1,0 @@
-.expire-enter {
-  opacity: 0;
-}
-.expire-enter-active {
-  opacity: 1;
-}
-.expire-exit {
-  opacity: 1;
-}
-.expire-exit-active {
-  opacity: 0;
-}
